Add File Browser.
This adds a file browser for viewing files on the EV3. It is limited to
browsing the home/ folder (since brickman runs as root). Directories are
indicated by a trailing '/' and executable files are indicated by a trailing
'*'. Selecting a directory will open it. Selecting an executable will run it
not as root, but as the user specified in the uid of the file. If a file is
owned by a system user (uid < 1000) then it will not be executed. Pressing and
holding the back button for more than 1 second will terminate a user program.