Add link to network connection info from bluetooth.
This connects the bluetooth controller back to the network controller so that
a bluetooth device that has a network connection can open the network connection
info window from the bluetooth device info window.

It also binds the powered checkbox on the main bluetooth window back to the
bluetooth technology in the network controller.