Add Wi-Fi menu under Networking.
This is like Bluetooth in that there is a list of Wi-fi devices. Selecting
a device opens a new window the displays info about the connection and allows
the user to connect or disconnect. There is also a "Forget" button to remove
credentials. And also like Bluetooth, there is a button that links to the
network connection.