Refactor IBrickManagerModule to allow lazy loading of main windows.
There is still more work to be done, but the idea is that windows should not
be created until they are requested and should be disposed when they are closed
to reduce the memory footprint of brickman. The lazy creation is partially
implemented here, but the disposal is not implemented at all yet.