openroberta: don't switch consoles
Disable console switching for openrobertalab to prevent lock up of brickman. This was partially done already, but it looks like we missed an instance.

Thre is a race condition where openrobertalab has probably already switched to graphics mode and so the system call to chvt hangs forever because it cannot switch consoles.

This is a temporary workaround to fix the lockup until a proper solution can be found. Currently, brickman will still be on the dispaly when openroberta programs are running. Ideally, the openroberta service should be updated to handle console switching to avoid the race condition.

Issue: OpenRoberta/robertalab-ev3dev#64